+++ example/src/index.js
@@ -1,15 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import { ThemeProvider } from 'styled-components';
-import { theme, Box } from 'krado-react';
-// import './styles/index.css';
+import { ThemeProvider } from 'theme-ui';
+import { theme } from 'krado-react';
 
 ReactDOM.render(
   <ThemeProvider theme={theme}>
-    <Box variant="root">
-      <App />
-    </Box>
+    <App />
   </ThemeProvider>,
   document.getElementById('root')
 );