+++ src/Avatar/Avatar.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Avatar as ThemeUiAvatar } from 'theme-ui';
+
+export default function Avatar({ ...rest }) {
+  return <ThemeUiAvatar {...rest} />;
+}