+++ src/Flex/Flex.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Flex as ThemeUiFlex } from 'theme-ui';
+
+export default function Flex({ ...rest }) {
+  return <ThemeUiFlex {...rest} />;
+}