+++ src/Grid/Grid.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Grid as ThemeUiGrid } from 'theme-ui';
+
+export default function Grid({ ...rest }) {
+  return <ThemeUiGrid {...rest} />;
+}