+++ src/index.js
@@ -1,18 +1,16 @@
-export { theme } from './theme';
+export { default as theme } from './theme';
 export { default as colors } from './colors';
 
 // Layout
-export { default as Box } from './Box';
-export { default as Flex } from './Flex';
-export { default as Grid } from './Grid';
+export { default as Box } from './Box/Box';
+export { default as Flex } from './Flex/Flex';
+export { default as Grid } from './Grid/Grid';
 
 // Components
-export { default as Text } from './Text';
-export { default as Heading } from './Heading';
-export { default as Title } from './Title';
-export { default as Link } from './Link';
-export { default as Card } from './Card';
-export { default as Image } from './Image';
-export { default as Avatar } from './Avatar';
-export { default as Button } from './Button';
-export { default as Tag } from './Tag';
+export { default as Text } from './Text/Text';
+export { default as Heading } from './Heading/Heading';
+export { default as Link } from './Link/Link';
+export { default as Card } from './Card/Card';
+export { default as Image } from './Image/Image';
+export { default as Avatar } from './Avatar/Avatar';
+export { default as Button } from './Button/Button';